[*] Refactoring
